server/requirements.txt
openenv-core[core]>=0.2.0
uvicorn>=0.24.0
fastapi>=0.115.0
pandas>=2.0.0
torch>=2.0.0
requests>=2.31.0
transformers>=4.40.0
accelerate>=0.30.0
trl>=0.9.0
peft>=0.11.0
datasets>=2.19.0
openai>=1.0.0